--- HowNet/.gitattributes
@@ -1,1 +1,0 @@
-HowNet_dict_complete filter=lfs diff=lfs merge=lfs -text
