+++ OpenHowNet/.gitattributes
@@ -1,0 +1,1 @@
+HowNet_dict_complete filter=lfs diff=lfs merge=lfs -text
