+++ OpenHowNet/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import absolute_import
+
+from OpenHowNet.Standards import HowNetDict
+from OpenHowNet.Download import download
+from OpenHowNet import pack
+
+name = "OpenHowNet"
