+++ OpenHowNet/sememe_all.txt
@@ -1,0 +1,1 @@
+模仿 保护 诱人性 问候 数量关系 仅 经济 捣乱 弊 劝说 宁乱 得罪 定期 能力值 辣 漂 友善性值 泰国 信息载体 之间 特殊性 爱尔兰 灰 幅度 不当 程度值 不稳 表示坏情感 构助 上海 发射 善待 注意 仪态 男 怯 卢旺达 中非 遭受 首次 制造 毛里求斯 扎 规矩 分析 状况 非 变形状 无礼 有序 浓 拔出 茂 疼痛 未成熟 比较 坐蹲 摩洛哥 暗 废 精 乌干达 足 朝向 采集 全面 使脏 巴布亚 平均 搀扶 不公正 推荐 积极 安哥拉 赠 叙利亚 使继续 逻辑性 印度尼西亚 支撑 托住 示喜 无望 甜 长时间 不说 准确性 叮 马达加斯加 借入 变空间位置 扎伊尔 文字 无 澳门 柔 好情 低植 气 拉脱维亚 抢 事务 贱 埋入 晚期 赞比亚 立场 格林纳达 利弊 处理 排泄 有关 清晰度 超 地方 组织 不重要 臭名 暴 印度 免除 堵塞 文莱 形成 独自 房间 喜悦 穿戴 胖瘦 标点 不讲理 利用 不适 长度值 丹麦 体格 犹豫 降级 集聚 荣 激怒 装束值 继续 巴勒斯坦 素 民用 质量 角度 萨摩亚 谴责 投下 黏度 凉 支 白俄罗斯 可能 相等 不足 放纵 罗得岛 前景差 前 相像 猜测 监察 赎 渺 出现 标准 锻炼 颠倒 浅薄 适合 无序 后天 加重 特定牌子 空满 正向 贞操 身分 喊 条约 少 塞舌尔 硬 怜悯 深度 按时 爬 辜负 明暗 南洋 上移 加快 气味 北朝鲜 难听 萨尔瓦多 升级 发炎 下去 防守 房屋 东京 络 攻打 不匀称 合并 必要性 问 已经 相关性 自杀 尝试 着重 比率 研究 量度 适应性 相比关系 吐出 英国 欧洲 原因 处于 帮助 不诚实 大概 征收 景象 经典 读物 相交 示思念 示不满 混合 精装 极 味淡 足球 辱 能 空 风格 主 有利 有创意 友善 取乐 票据 特性 中美洲 年龄 粗细值 缘于 加浓 下 疏远 乞求 高尚 脏 软 跟随 当作 相关 淘 滚 倒霉 分别的 衔起 明显性 好听 薄 瞒 周围 浸入 喜欢 慷 苏里南 费用 凹 证券 控制 深浅 距离值 洞察力 弯度 巴布亚新几内亚 逃跑 和谐 体育 木 偷 建立 易变 阿富汗 温 反抗 增多 冒险 证明 已婚 先进 连接 抗议 羞愧 栽植 序数 失望 物形 凸 衰败 感受 可靠性 逆 柴可夫斯基 理智 身矮 长 被动 久暂值 胡闹 来源值 匀称性 离开 生物 无疵 不同 新加坡 贵 瘫麻 抚 肯尼亚 愚 赞成 饮品 拾起 放松 抽搐 总计 解密 整 变 科摩罗 答 忠 语言 否 出去 珍 收 先进性 好看 荒 程度 花草 建造 尼加拉瓜 深刻 衣物 问世 回移 时间特性 莽 无能 开除 秋 地力 活动 包扎 早期 偷懒 方向性他移 划定 久暂 补做 考试 点 扭转 不舍 典雅 落落大方 方位特性值 速度值 左 状态 致敬 延续 感动 频度值 严重性 诱人 性能 给 当 格鲁吉亚 谋取 促使 概数 好坏 不必要 爱惜 妒忌 单位 拿 香 波利尼西亚 稍 捷克和斯洛伐克 成熟 招惹 出版 独特 加纳 仁 消闲 搂抱 嫁 著名 无生命 隐秘 韩国 无创意 习惯性 低等 主次 外移 分钟 外包装值 指代 功过 赚 诊察 荣辱 伊朗 排泄物 慢 画 本土 荷兰 射击 冷清 削减 使之动 住下 旬 相信 发送 纯化 胆量 做梦 不正常 成熟度 布隆迪 惬 埋怨 尾 陡 跑 便 日本 拴连 嗜好 咀嚼 制冷 差遣 功率 国籍值 留给 经受 总是 失去 战胜 不一致 未婚 没志气 缓急 时间特性值 斯威士兰 洁净 从不 清醒 看 饲养 元首 使锋利 动果 通俗 越南 机密 安慰 窄 疑似 非洲 乱 不挤 特立尼达和多巴哥 擦拭 讲和 面 教养 认知 它 乐器 尖 翅 板球 杂 热情 高矮 马拉维 中东 提及 强过 不容 行动 尼日利亚 可靠 坏风气 蛮干 使存现 意大利 争斗 迟 消灭 阿曼 总 欺骗 等待 黏 代谢 合逻辑 淫 僧伽罗 交往 特别 婉 描写 引用 印刷 普通 部件 肿胀 搬动 携带 泛动 长度 归属地方 与女性相关 哥斯达黎加 供 气味值 能力 间接 使他人感知 审问 喝 小辈 安 道歉 半 傲气 抄写 吉布提 惋惜 干 修理 庸 爱斯基摩 分发 满意 最佳 归属中央 良态 翘耸 面积 右 巴林 精确 醒 赌博 示怒 秘鲁 代替 漏出 酒 保加利亚 菲律宾 篮球 兴衰 侧 表示 火 病态 无价值 垒球 祝贺 加害 谈话 免费 吓唬 东帝汶 机动性 无诚心 具体 光 共 莱索托 高 拟人 多 农 多余 独立 颤动 苏格兰 自私 管理 通俗性 经常 心 量变 临时 成形 秃 灿烂 交配 均衡 实体 勤 外来 敢于 付 回想 内容值 散 打 永久性 间断 小 可信 高度 削弱 鼓吹 背叛 手 同意 会 减慢 非常好 积极性 还 充实 哭泣 消化 歪 有可能 海地 塞浦路斯 变外观 有罪 走兽 欠 自由度 禽 壮观 享受 加蓬 聚集 强 同辈 吃饱 烹调 推 反面 保险 转动 正常 现象 招待 胚 境况值 洒下 紫 驱赶 离别 法国 重复 有毒 阿拉伯联合酋长国 心地 装载 频度 波斯 他 保持 功 植物 生物学 直 正误 精神 没羞 属于 广 勾引 有能力 使净 缺 大 香港 人工物 过去 拆卸 洁 不坚决 南美 长辈 插入 液 斯里兰卡 揭露 死 民事 复 友 刚果 心不在焉 上去 线 配偶 接受 使成长 摆动 价格 借出 态度 唱 休息 运送 物理状态 好运 形状值 境况 龊 热量 厚 国家 上次 适宜 展示 实验 沙特阿拉伯 衰变 粗 全额 古巴 缠绕 情绪 指使 偏僻 有准备 切削 乌克兰 乡 非法 身 龄大 器具 注意力 坏天 俄罗斯 合作 智利 样式值 忍耐 理睬 很 标志 吉尔吉斯斯坦 拐弯 龄小 到达 伊斯兰教 临产 思想 谋生 有形性 象牙海岸 无意 美丑 芬兰 实现 纪律性 静态 汤加 娶 加湿 评估 回避 规划 屈服 斯洛文尼亚 推理 包围 界限 上 有功劳 志气 辫编 喷出 浇灌 罗得西亚 呼出 诚心 接近 怀孕 诚实 尖声 不友善 疲乏 疏通 合作性 劫难 礼貌 放置 智慧 价值值 私心 进入 木然 货币 争吵 领会 食物 表示情感 多少 下移 不相关 国都 同步 关系 苏丹 邀请 相互 无色 赊 容积 常态 天象 正面 称作 引导 为难 音质值 有备性 阿联酋 疯痴 内 创造 难 啃咬 稳定性 厚度 变特性 不满 分离 也门 丧偶 基数 缅甸 新近 齐 快 不成形 动物 命令 嵌入 军 经过 道路 仇恨 黄 创造性 深 表情 冰 维吾尔 搭乘 与男性相关 健康 选拔 亲密 柬埔寨 晨 调查 结果 交 墨西哥 敌 澳洲 取 单 工 异 崎岖 易 追赶 场所 浑 偶尔 己 示不同意 青 臂 低声 南斯拉夫 异步 恒定 保存 花费 谦 北京 困扰 斐济 音乐 繁 嫩 约旦 坚决 达荷美 来源 强迫 召唤 遵循 埃及 表演 粗细 符号 归属值 突尼斯 商讨 微生物 自移 药物 方向 习惯值 渐进性 导致 县 加工状况 冲淡 始 外观 扣住 奥地利 优化 忧愁 占领 致哀 羡慕 仓促 方法 狭 船 身分值 冷漠 端 学问 台湾 不值 马里 味道 难易 普鲁士 存在 直接 犹太人 终生 没可能 直系 测量 自愿 问题 道理 体液 靠近 阿尔及利亚 价格值 叠 疾病 动助 举止 消极 信息元素 朝代 正经 发达 液态 装饰 使误知 讲求 傲 照射 关联 沃 犹太教 棒球 道德 边 扔 夸奖 制止 旋即 承担 着急 扩大 一般性 有耐性 泛 使喜悦 加工状况值 对待 使动 挖掘 商业 大洋洲 真伪 共通 季 干燥 稀 中年 混淆 距离 社交性 坦桑尼亚 淡 团体 抓挠 彩 昏迷 北大西洋 灵 锐利度 此刻 脚 红 流浪 不宜 赐 朝鲜 冬 现场 天变 新 毛里塔尼亚 辈分 值得 勾结 多哥 政 劣质 黑 中国 索马里 比拟 纤 不好用 好天 辩论 完毕 取悦 格陵兰 南极洲 阿塞拜疆 宁 性别值 塞尔维亚 矿 致谢 性别 缺少 权利 姓 未加工 纳入 分类 俗气 亏损 阵发性 卖力 灭亡 起始 数量 牙买加 直率性 生殖 黎巴嫩 浓淡 拥护 整洁 类型 爱 新旧 补偿 天主教 羽毛球 事件 样式 乐观 最后 西 润滑 无准备 有味 假装 族群 任命 告诉 外观变 预料 合理 呆 多种 兴 旁系 成长 剩余 次序 寻欢 中等 狡 安装 资金 莫扎特 破开 勇 化学 物理 整缺 倒出 因特网 功能词 必要 雷 表现 非正统 安心 音质 莫桑比克 挪威 起身 连续 正义性 研磨 翻译 稠 健康值 绿 性能值 示爱 金属 押给 外观值 折弯 省 剥夺 上沃尔特 方 深奥 波多黎各 夸大 无罪 戒除 终结 使之非 希腊 克罗地亚 净额 使不变 不合格 有纪律 厄瓜多尔 隶属 数量值 以色列 春 无毒 铺摊 已加工 制度 后裔 打结 光洁度 成绩 承认 罢免 辈分值 蓝 设施 家养 使相等 无纪律 浅 卸载 先天 正义 拉直 涂抹 衰 沙哑 密度 基于 松紧 成名 不雅 无知 违背 高声 捷克 纵 不可靠 诵读 越过 孤 提出 活着 奢 无形 尼泊尔 职位 悬挂 成年 畸 僻 美 他移 欢迎 人为 动趋 不贪心 正反 影响 幸福 恭顺 口 配件 德国 飞行器 改正 趣 雅 痕 时间 深浅值 无关 没学问 不会 动量 吸入 去 禁止 陋 改变 南非 充实性 咸 控告 瑞士 瘠 轻视 终止他移 无功劳 决定 水 庄 自由 形变 复杂度 甘比亚 未定 感知状态 毒品 阿根廷 巴拉圭 伪 牲畜 填入 俭 站立 贪心 适当性 天然物 变感知 根 糊涂 婚姻状况 阿拉伯 是 泼冷水 简 说 眼 知道 捐 停做 巴哈马 遮盖 变性态 背令 虫 巴尔干半岛 生活状况 基督教 成为 卖 词语 买 体 降低 旧 苏联 熟悉性 美化 不知 全 丑 怀疑 风气 头 佛得角 一致 念头 摸 蒙古 区 传送 某 不精明 沉着 属性 宽度 酬金 拍摄 睡 圆 天然 刑事 难吃 中位 显现 不健 将来 表示好情感 心怀 厌恶 也 马其顿 邮寄 担任 贞洁 消失 到期 保养 非正义 重量 数学 没条理 敬佩 圭亚那 效果值 发誓 等级值 腿 摩擦 颜色值 关闭 重要性 和谐性 秒 舔 埃塞俄比亚 能够 严厉 出血 变领属 合法性 弄懂 提早 体积 国际 外国 无私 克里米亚 运动器材 慎 拒做 电 倾斜 北美 葡萄牙 叹气 等级 笑 输掉 深度值 挤出 弱 音量值 认为 怪 飞 扁 不同意 拖延 南 娱乐 态变 有道德 诉讼 照料 好风气 效果差 有教养 军用 圣多美和普林西比 巧 纳米比亚 月 矫揉造作 匀称 明 溢出 打开 定住 吉祥 内容 公开 电阻 拦住 跳 末 责任 糙 合格 利 世界 部分 机敏性 冷 悲惨 巴西 哥伦比亚 水域 趣味 安危 感激 坏脾气 取回 票证 细心 片面 原谅 折断 文书 贬低 稳固 亲疏值 开关 塞拉利昂 形状规则性 匈牙利 一致性 科威特 锡金 倔强 焚烧 玻利维亚 忘记 西藏 灰心 喂 供养 守时性 几内亚 结盟 饥渴 外交 巴斯克 穷 悲哀 牢 有礼 洪都拉斯 形状规则 加热 费力 不道德 平坦性 群体 静 正 化妆 赤裸 不做 寒热 赭 宗教 金融 失败 安道尔 挤 坡度 普通性 后退 懒 卢森堡 教 使之是 特定 肉 提高 未达成 圣马力诺 密级 夜 有弊 豪华 官 礼貌值 归属 树 土气 周 有用 罗马尼亚 变颜色 皇 急切 低于正常 紧 过程 肢 专 湿度 土耳其 正确 做 白 按压 家庭 听 琐 云雾 滤清 捷 忙 大地 流放 即将 分辨 感情特性 天体 簿册 不直率 午 乏 盘旋 举止值 变量度 维值 侦察 不合逻辑 有志气 伊拉克 度过 不便 电量 调解 从事 生理学 巴巴多斯 网 卑劣 节 拉 高等 褐 苦 骨 坏掉 清 清浑 平装 感知 方位特性 暴动 中 单弱 直率 位置 发表 比利时 发声 危地马拉 回应 有知 教养值 味浓 收受 夏 武器 闹 达成 形状不规则 试试 抗住 成功 否认 内移 实例 吸收 点燃 讲理 利比亚 陪衬 康健 排出 歹 次 不精确 要求 远 快干 触 脆 准确性值 西萨摩亚 不熟悉 气态 含混 异同 健壮 下次 机构 前往 声 宽大 女 游 杀害 调整 使活 弯 时 说明 爱沙尼亚 忠孝 几乎 诽谤 不丹 马尔代夫 驾驭 不善交往 平 密 乒乓球 危 合理性 勒紧 跌倒 险 涝 专注 相同 医治 姿势 几内亚比绍 重 塔吉克斯坦 传染 练习 条理性 完结 计算 没能力 倚靠 价值 太平洋 西班牙 喀麦隆 弛 不安 炫耀 近 恐怖袭击 多少值 意愿 正宗性 部件自移 外表值 误信 缓 强制性 稳固性 人 国籍 整理 题写 骤然 亲疏 图像 详 道教 损耗 优质 加深 雕刻 镇静 律法 耐性 用具 角 年龄值 多米尼加 对立 算术 得到 场面 珍宝 白昼 前进 吸引 不忠 有威力 吝 没教养 兽 编辑 物质 非正式 波兰 属性值 效果 前景佳 语文 从容 妖媚 查 摩纳哥 增加 巴比伦 亚洲 相距 有效 初等 易于 书刊 躺卧 交流 皮 身高 害怕 辉煌 变级 方式性自移 生命 传播 桥梁 现在 表演物 躲避 特性值 化学物 意志 回绝 恋爱 返回 功用 智 神气 大西洋 倔劲 你 满 好用 残疾 裁定 友善性 投靠 血统 味道值 方向性自移 利比里亚 风度 力量 水果 反向 寻 合格性 落后 比赛 定向 有形 救助 悲观 共同 有 废物 纯 偶发 有情 陆地 协调 空洞 约定 疏 退出 瘦 发端 讷 无味 加固 会面 出错 情感 出身 受伤 邋遢 知识 活 缩小 嗜好物 年 止动 浪费 可 尼日尔 拳击 清晰 奖励 正反值 温度 语助 艰苦性 烟草 有条理 文 安排 宽宏性 状况值 原 有学问 电压值 存现 撒 排球 动量值 疑问 名量 熟悉 发痒 马来西亚 洗涤 写 空间 留存 鱼 百慕大 后 连续性 有望 依靠 粗心 无生物 培养 选择 哈萨克斯坦 固态 罪恶 较 强度 感情特性值 公开性 期望 阻止 湿 低 渐进 日 间 越界 功用值 谨慎性 建筑物 网球 新闻 风 准备 过 计数 电压 通往 无效果 感叹 踢踩 市 音量 俗 前景 正式 来 时髦 阿尔巴尼亚 思念 离婚 美洲 类型值 包含 心胸狭窄 减少 抽象性 机器 瑞典 意图性 不贞 土石 雨雪 流 假定 交通工具 便利 佛教 编号 变属性 蔬菜 反对 姓名 坚定性 严格 善交往 捞起 空域 改良 节省 老挝 露出 马尔他 巴基斯坦 姓名值 立陶宛 委内瑞拉 驶 背起 津巴布韦 祖先 骗 姿势值 浮 嘈杂 吃 不愿 慢待 难看 尚未 循环 乌拉圭 惊奇 斯堪的纳维亚 部件他移 贝宁 停留 非自动 臭 纷乱 钱财 不诱人 正统 罹患 记得 信息 加拿大 雇用 命运 抱歉 提升 速度 好吃 教育 冷峻 释放 与特定国家相关 真 严重 启动 方向特性值 事情 误 列支敦士登 扼要 实施 旅游 心胸开阔 斯洛伐克 肖邦 示同意 高于正常 脱俗 短时间 计划 消除 闻 灵巧性 赤道几内亚 巴拿马 反驳 无愧 分歧 委托 私 脱离 装束 松 保证 气喘 情况 目的 围绕 乌兹别克斯坦 腹症 电脑 所罗门群岛 烦恼 外包装 些 车 博茨瓦纳 感觉 处罚 生气 北 伟 霸道 特意 波希米亚 孟加拉国 浓度 尺寸 土库曼斯坦 捉住 皱 重要 野生 胖 职位值 形状 交换 获胜 卡塔尔 有声 懊悔 损害 恢复 艺 进展 闲 硬度 急 不和谐 材料 剥去 畅 旱 警 走 复原 公正性 藏匿 早 钝 乍得 已定 撞 维 不祥 塞内加尔 高贵 纯正度 横 无声 贫富 澳大利亚 变是非 尺寸值 坏情 形式 需求 揉弄 医 纽约 伪造 气量 食品 瑙鲁 冰岛 万物 庄稼 无用 瞎说 富 势力 众 不凡 咽下 少儿 老年 亚美尼亚 散开 摇动 名声 开始 家具 较量 酸 舍得 发生 滑 幅度值 摺叠 有生命 信件 部 学习 看望 韧 再 毛 激动 另 受限 放弃 似 抽象 美国 公正 最 知网 过值 东 不合理 呼吸 轻微 习惯 新西兰 无威力 读 青年 结婚 合法 不如 阔 居民区 拉丁美洲 颜色 军队 无名 嘲讽 闷 软件 热 轻 报复 记录 集体的 变亮度 短 外 不费力 贝多芬 愿意 思考 来回 自动 我 联合国
